refactor(theme): extract config validation
